refactor: refactor after review moving functions
