+++ src/app/(dashboard)/facteurs-d-emission/[id]/modifier/page.tsx
@@ -2,8 +2,8 @@ import withAuth from '@/components/hoc/withAuth'
 import EditEmissionFactorPage from '@/components/pages/EditEmissionFactor'
 import NotFound from '@/components/pages/NotFound'
 import { canEditEmissionFactor } from '@/services/permissions/emissionFactor'
+import { hasAccessToEmissionFactor } from '@/services/permissions/environment'
 import { getDetailedEmissionFactor } from '@/services/serverFunctions/emissionFactor'
-import { hasAccessToEmissionFactor } from '@/utils/permissions'
 import { UserSession } from 'next-auth'
 import { redirect } from 'next/navigation'
