+++ src/components/home/UserView.tsx
@@ -2,8 +2,8 @@ import { getAccountOrganizationVersions } from '@/db/account'
 import { OrganizationVersionWithOrganization } from '@/db/organization'
 import { hasAccountToValidateInOrganization } from '@/db/user'
 import { default as CUTLogosHome } from '@/environments/cut/home/LogosHome'
+import { hasAccessToActualityCards } from '@/services/permissions/environment'
 import { canEditMemberRole } from '@/utils/organization'
-import { hasAccessToActualityCards } from '@/utils/permissions'
 import { UserSession } from 'next-auth'
 import ActualitiesCards from '../actuality/ActualitiesCards'
 import Onboarding from '../onboarding/Onboarding'
