+++ src/environments/cut/study/StudyHomeMessage.tsx
@@ -1,6 +1,5 @@
 'use client'
 import { UserSessionProps } from '@/components/hoc/withAuth'
-import { hasAccessToEnvironment } from '@/utils/userAccounts'
 import { Alert, List, ListItem, ListItemText, Typography } from '@mui/material'
 import { Environment } from '@prisma/client'
 import { useTranslations } from 'next-intl'
@@ -8,7 +7,7 @@ import { useMemo } from 'react'
 import styles from './StudyHomeMessage.module.css'
 
 const StudyHomeMessage = ({ user }: UserSessionProps) => {
-  const isCut = useMemo(() => hasAccessToEnvironment(user, Environment.CUT), [user?.environment])
+  const isCut = useMemo(() => user.environment === Environment.CUT, [user?.environment])
 
   const t = useTranslations('home.message')
   return isCut ? (
